evolve-backend/requirements.txt
Django
djangorestframework
dj-database-url
gunicorn
psycopg2-binary
whitenoise
python-decouple
python-dotenv
requests
twilio
Pillow
PyJWT
urllib3
certifi
charset-normalizer
idna
sqlparse
typing_extensions
asgiref
packaging
python-monkey-business
django-nested-admin  
rest_framework_simplejwt
djangorestframework-simplejwt
python-docx
django-phonenumber-field[phonenumbers]
django-filter
# django-otp
# qrcode
django-compressor
django-redis
groq
beautifulsoup4>=4.12
selenium>=4.19
webdriver-manager>=4.0
tqdm>=4.66
django-vite>=3.1.0
pydantic
pydantic-core